+++ src/main/java/br/pucrio/les/esg_token_backend/configurations/security/SecurityConfiguration.java
@@ -1,0 +1,28 @@
+package br.pucrio.les.esg_token_backend.configurations.security;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@EnableWebSecurity
+@Configuration
+public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+    // Configurations for authentication
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+    }
+
+    // Configuration for authorization
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+    }
+
+    // Configuration for static resources
+    @Override
+    public void configure(WebSecurity web) throws Exception {
+
+    }
+}